requirements.txt
requests==2.22.0
python-dotenv==0.12.0
vk-api==11.8.0
PySocks==1.7.1
python-telegram-bot==12.5.1
google-api-python-client==1.9.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
PyDrive==1.3.1
urllib3==1.25.8